// src/Enum.ts
export enum BillChoice {
    CREATE_BILL = 1,
    SHOW_LIST_BILL = 2,
    UPDATE_BILL = 3,
    DELETE_BILL = 4,
    FIND_BILL = 5,
    SOFT_BILL = 6,
    Amout = 7,
    EXIT = 0
}

export enum UpdateChoice {
    UPDATE_ALL_INFO_BILL = 1,
    UPDATE_NAME = 2,
    UPDATE_HOMENUMBER = 3,
    UPDATE_ELECTRICMETERID = 4,
    UPDATE_OLDINDEX = 5,
    UPDATE_NEWINDEX = 6,
    EXIT = 0
}